volatile interface option, config name on the command line
Because my fedora server  switched to cgroups v2, the docker stopped to work and I was forced to runn HomeAssistant under podman, which apparently does just fine.
So the docker0 and hassio interfaces are gone, but I wanted to save my settings in case of switching back to docker again in the nearest future. So I added options->volatile to teh interface to indicate that it can be non-existent at the generation phase. so no sanity checks are done, but iptables process it w/o hesitance.